post test addded

+++ src/modules/post/post.services.ts
@@ -1,13 +1,24 @@
 import AppError from "../../errors/appError";
 import { status } from "../../utils/status";
-import { Prisma } from "../../../generated/prisma/client";
 import PostRepository from "./post.repository";
+import {
+  CreatePostDto,
+  UpdatePostDto,
+} from "./post.interface";
 
 class PostService {
   constructor(private postRepo: PostRepository) {}
 
-  async createPost(data: Prisma.PostCreateInput) {
-    return this.postRepo.createPost(data);
+  async createPost(data: CreatePostDto) {
+    return this.postRepo.createPost({
+      title: data.title,
+      content: data.content ?? null,
+      author: {
+        connect: {
+          id: data.authorId,
+        },
+      },
+    });
   }
 
   async findAllPosts(query: Record<string, unknown>) {
@@ -16,19 +27,61 @@ class PostService {
 
   async findPostById(id: string) {
     if (!id) {
-      throw new AppError("Enter a valid post Id", status.NOT_FOUND);
+      throw new AppError(
+        "Enter a valid post id.",
+        status.BAD_REQUEST
+      );
     }
-    return this.postRepo.findPostById(id);
+
+    const post = await this.postRepo.findPostById(id);
+
+    if (!post) {
+      throw new AppError(
+        "Post not found.",
+        status.NOT_FOUND
+      );
+    }
+
+    return post;
   }
 
-  async updatePost(id: string, data: Prisma.PostUpdateInput) {
+  async updatePost(id: string, data: UpdatePostDto) {
     if (!id) {
-      throw new AppError("Enter a valid post Id", status.NOT_FOUND);
+      throw new AppError(
+        "Enter a valid post id.",
+        status.BAD_REQUEST
+      );
     }
+
+    const post = await this.postRepo.findPostById(id);
+
+    if (!post) {
+      throw new AppError(
+        "Post not found.",
+        status.NOT_FOUND
+      );
+    }
+
     return this.postRepo.updatePost(id, data);
   }
 
   async deletePost(id: string) {
+    if (!id) {
+      throw new AppError(
+        "Enter a valid post id.",
+        status.BAD_REQUEST
+      );
+    }
+
+    const post = await this.postRepo.findPostById(id);
+
+    if (!post) {
+      throw new AppError(
+        "Post not found.",
+        status.NOT_FOUND
+      );
+    }
+
     return this.postRepo.deletePost(id);
   }
 }
